examples/classification/requirements.txt
argcomplete==1.12.1
avro-python3==1.9.2.1
azure-storage-blob==12.4.0
bottle==0.12.19
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.4
click==8.0.1
crcmod==1.7
cycler==0.10.0
diffimg==0.2.3
docopt==0.6.2
fastavro==1.4.1
filelock==3.0.12
fire
fusepy==2.0.4
future==0.18.2
gdown==3.13.0
httplib2==0.17.4
idna==3.2
imageio==2.9.0
indexed-gzip-fileobj-fork-epicfaace==1.5.4
isodate==0.6.0
joblib==1.0.1
kiwisolver==1.3.1
markdown2==2.3.10
marshmallow==2.15.1
marshmallow-jsonapi==0.15.1
matplotlib==3.4.3
mock==2.0.0
networkx==2.6.2
nltk==3.6.2
numpy==1.21.2
oauth2client==4.1.3
opacus==0.13.0
packaging==21.0
pandas==1.3.2
pathtools==0.1.2
pbr==5.6.0
Pillow==8.3.1
psutil==5.7.2
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.2
pymongo==3.11.4
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.*
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.*
regex==2021.8.3
requests
retry==0.9.2
sacremoses==0.0.45
scikit-image==0.18.2
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.2
selenium==3.141.0
sentence-transformers>=2.0.0
sentencepiece==0.1.96
sentry-sdk==0.18.0
six==1.15.0
SQLAlchemy==1.3.19
termcolor==1.1.0
threadpoolctl==2.2.0
tifffile==2021.8.8
tokenizers==0.10.3
tqdm>=4.62.1
typing-extensions==3.7.4.3
urllib3==1.26.*
watchdog==0.10.3
websocket-client==1.0.1
git+git://github.com/lxuechen/swissknife.git
